Apply new source structure to abort.c
