+++ libraries/ah_base/src/abort.c
@@ -8,40 +8,8 @@
 
 #include "ah/defs.h"
 
-#include <signal.h>
 #include <stdarg.h>
 #include <stdio.h>
-
-#if AH_IS_WIN32
-#    define WIN32_LEAN_AND_MEAN
-#    include <windows.h>
-#endif
-
-#if AH_HAS_POSIX
-#    include <unistd.h>
-#else
-#    include <signal.h>
-#endif
-
-ah_extern void ah_abort()
-{
-#if AH_HAS_POSIX
-
-    struct sigaction act;
-    (void) sigemptyset(&act.sa_mask);
-    (void) sigaddset(&act.sa_mask, SIGABRT);
-    (void) sigprocmask(SIG_UNBLOCK, &act.sa_mask, NULL);
-
-    (void) kill(getpid(), SIGABRT);
-
-#else
-
-    raise(SIGABRT);
-
-#endif
-
-    ah_trap();
-}
 
 ah_extern ah_noreturn void ah_abortf(const char* format, ...)
 {
